BC-1502 - code review
